+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Badges: `[UPDATED]`, `[FIXED]`, `[NEW]`, `[DEPRECATED]`, `[REMOVED]`,  `[BREAKIN
 * `[NEW]` - uses `org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4`
 * `[NEW][Experimental]` - Background module loading thanks to Kotlin coroutines with `lazyModules()`, `asyncModules()` KoinApplication DSL functions. `awaitKoinStart()` suspend function to wait background completetion. JVM `waitKoinStart()` function is using runBlocking to wait for completion.
 * `[NEW][Experimental]` - Koin Lazy Module, as a Module that don't not trigger any resources preparation. The `lazyModule { }` function help declare such `Lazy<Module>`
+* `[NEW][Experimental]` - API functions to run Koin function after start: `runOnKoinStarted` and `onKoinStarted`
 
 ## [test-3.4.0]() - 2023-03-24
 * `[UPDATED]`- update core-3.4.0